add validation error message in add package

+++ app/model/package.go
@@ -12,7 +12,7 @@ type Package struct {
 func (p Package) Validate()(error, map[string]string) {
 	return p.Model.ValidationRules(&p, 
 	validation.Field(&p.Description, validation.Required.Error("Description is required."), validation.Length(1, 0).Error("Description is required.")), 
-	validation.Field(&p.Price, validation.Required.Error("Price is required."), validation.Min(1).Error("Price is required.")), )
+	validation.Field(&p.Price, validation.Required.Error("Price is required."), validation.Min(float64(1)).Error("Price is required.")), )
 }
